bug fix) exception parse number, etf price
2026-05-04 13:13:15 [ERROR] j.CrawlPriceLatestDataGoKrEtfJob :142 ◁
┝ Exception:: For input string: "3672903674"
java.lang.NumberFormatException: For input string: "3672903674"
        at java.base/java.lang.NumberFormatException.forInputString(NumberFormatException.java:67)
        at java.base/java.lang.Integer.parseInt(Integer.java:662)
        at java.base/java.lang.Integer.parseInt(Integer.java:778)
        at kr.andold.stock.service.DataGoKrService.toPriceDomain(DataGoKrService.java:78)
        at kr.andold.stock.job.CrawlPriceLatestDataGoKrEtfJob.main(CrawlPriceLatestDataGoKrEtfJob.java:122)
        at kr.andold.stock.job.CrawlPriceLatestDataGoKrEtfJob.call(CrawlPriceLatestDataGoKrEtfJob.java:43)
        at kr.andold.stock.job.CrawlPriceLatestDataGoKrEtfJob.call(CrawlPriceLatestDataGoKrEtfJob.java:25)
        at java.base/java.util.concurrent.FutureTask.run(FutureTask.java:317)
        at java.base/java.util.concurrent.ThreadPoolExecutor.runWorker(ThreadPoolExecutor.java:1144)
        at java.base/java.util.concurrent.ThreadPoolExecutor$Worker.run(ThreadPoolExecutor.java:642)
        at java.base/java.lang.Thread.run(Thread.java:1583)
2026-05-04 13:13:15 [DEBUG] j.CrawlPriceLatestDataGoKrEtfJob :145 ◁
┕ 『EXCEPTION』 최근주가ETF::main() - 1953 ms

+++ src/main/java/kr/andold/stock/domain/ResultDataGoKr.java
@@ -141,7 +141,7 @@ public class ResultDataGoKr<T> {
 		private Integer mkp;
 		private Integer hipr;
 		private Integer lopr;
-		private Integer trqu;
+		private Long trqu;
 		private String trPrc;
 		private String lstgStCnt;
 		private String mrktTotAmt;
